chore : nickname Api mapping 변경
